+++ js/hdelk.js
@@ -53,7 +53,8 @@ var hdelk = (function(){
     var edge_color = '#888';
     var edge_highlight_color = ['#DDD', '#444', '#06d', '#C00', '#980', '#590', '#C40070', '#AD60FF', '#7579FF', '#a1a1a1', '#7fb6ee', '#e57f7f', '#ccc27f', '#adcc7f', '#e17fb9', '#d5b0ff', '#babdff'];
     var edge_highlight_width = 2;
-    var edge_bus_width = 6;
+    const DEFAULT_EDGE_BUS_VISUAL_WIDTH = 6;
+    var g_edge_bus_visual_width = DEFAULT_EDGE_BUS_VISUAL_WIDTH;
     var edge_bus_color = '#AAA';
     var edge_bus_highlight_color = ['#DDD', '#444', '#06d', '#C00', '#980', '#590', '#C40070', '#AD60FF', '#7579FF', '#a1a1a1', '#7fb6ee', '#e57f7f', '#ccc27f', '#adcc7f', '#e17fb9', '#d5b0ff', '#babdff'];
     var edge_bus_highlight_width = 6;
@@ -65,9 +66,11 @@ var hdelk = (function(){
      * Creates an SVG diagram from a JSON description.
      * @param {object} graph
      * @param {string} divname
+     * @param {number} edge_bus_visual_width Override the default edge bus width.
      * @param {number} elk_thoroughness How much effort should be spent to produce a nice layout. min=1, default=32
      */
-    var layout = function(graph, divname, elk_thoroughness=32) {
+    var layout = function(graph, divname, edge_bus_visual_width = DEFAULT_EDGE_BUS_VISUAL_WIDTH, elk_thoroughness=32) {
+        g_edge_bus_visual_width = edge_bus_visual_width;
         const elk = new ELK({})
 
         // create a dummy drawing just to get text sizes
@@ -702,7 +705,7 @@ var hdelk = (function(){
             }
         } else {
             if ( edge.bus ) {
-                width = edge_bus_width;
+                width = g_edge_bus_visual_width;
                 color = edge_bus_color[ edge.highlight ];
             } else {
                 width = edge_width;
@@ -710,7 +713,7 @@ var hdelk = (function(){
             }
         }
 
-        //         var width = ( edge.bus ) ? ( edge.highlight ?  edge_bus_highlight_width : edge_bus_width ) : ( edge.highlight ? edge_highlight_width : edge_width );
+        //         var width = ( edge.bus ) ? ( edge.highlight ?  edge_bus_highlight_width : g_edge_bus_visual_width ) : ( edge.highlight ? edge_highlight_width : edge_width );
         // var color = ( edge.bus ) ? ( edge.highlight ? edge_bus_highlight_color[ edge.highlight ] : edge_bus_color ): ( edge.highlight ? edge_highlight_color[ edge.highlight ] : edge_color );
 
         if ( sections ) {